Refactoring pakietów
