+++ src/main/java/pl/hycom/jira/plugins/gitlab/integration/service/processors/IssueCommitChangeNotificationProcessor.java
@@ -25,8 +25,8 @@ import com.atlassian.jira.user.ApplicationUser;
 //import com.atlassian.jira.user.DelegatingApplicationUser;
 import com.atlassian.jira.user.MockApplicationUser;
 import org.springframework.beans.factory.annotation.Autowired;
-import pl.hycom.jira.plugins.gitlab.integration.gitpanel.model.CommitData;
-import pl.hycom.jira.plugins.gitlab.integration.gitpanel.service.ProcessorInterface;
+import pl.hycom.jira.plugins.gitlab.integration.model.CommitData;
+import pl.hycom.jira.plugins.gitlab.integration.service.processors.ProcessorInterface;
 
 import java.util.HashMap;
 import java.util.Map;
